add: integration tests and jest mock for different describe
