+++ libs/nft/thirdweb-organizer/project.json
@@ -44,7 +44,8 @@
       "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
       "options": {
         "jestConfig": "libs/nft/thirdweb-organizer/jest.integration.ts",
-        "passWithNoTests": true
+        "passWithNoTests": true,
+        "runInBand": true
       },
       "configurations": {
         "ci": {
